Add files via upload
Triggered with a shortcode to show the order formula. Basicly easy ordering for customers that place large quantities to several customer at the same time. 
This plugin do not take into consideration stock levels, and that is with purpose since it is aimed at b2b customers that accepts delivery times and will have a confirmation of the orders.